+++ lib/hyde/helpers.rb
@@ -2,7 +2,7 @@ class Hyde
 module Helpers
   def partial(path, locals={})
     partial = Partial[path.to_s, page]  or return ''
-    partial.to_html :page => self
+    partial.to_html locals.merge(:page => self)
   end
 
   def rel(path)